+++ posttroll/backends/zmq/__init__.py
@@ -1,6 +1,24 @@
+import logging
+import os
+
 import zmq
 
 from posttroll import config
+
+logger = logging.getLogger(__name__)
+context = {}
+
+
+def get_context():
+    """Provide the context to use.
+
+    This function takes care of creating new contexts in case of forks.
+    """
+    pid = os.getpid()
+    if pid not in context:
+        context[pid] = zmq.Context()
+        logger.debug("renewed context for PID %d", pid)
+    return context[pid]
 
 def _set_tcp_keepalive(socket):
     _set_int_sockopt(socket, zmq.TCP_KEEPALIVE, config.get("tcp_keepalive", None))
